mysql change to local

+++ springboot-maven-multi/common/src/main/java/com/dubbo/service/DubboService.java
@@ -1,5 +1,5 @@
 package com.dubbo.service;
 
 public interface DubboService {
-    boolean login(Long id ,String password);
+    boolean login(int id ,String password);
 }
